Churn Analysis/requirements.txt
pandas
faker
numpy
matplotlib
seaborn
scikit-learn
arabic-reshaper
python-bidi